Fix condition failure when no devices are selected
The condition to check if the water_heater or other_devices are selected
was failing when no devices were selected. This was due to the fact that
the condition was checking the length of something that could be None.

This was causing the following errors:
Error: Template rendered invalid entity IDs: [None]

Linting the YAML file.